+++ NEWS.md
@@ -1,3 +1,11 @@
+# Package update 0.1.4 -> 0.1.5
+
+## New functions: 
+* New functions for price imputations (*data_imputing*) and for detecting new and disappearing products (*products* and *products_figure*).
+
+## Improvements: 
+* The **data_preparing** function has now two additional parameters which allow to consider zero prices and zero quantities.
+
 # Package update 0.1.3 -> 0.1.4
 
 ## New functions